Fix vario altitude value & display
Those were copies of the GPS altitude calculation. But GPS altitude
was stored in a 32 bit int, where vario values are stored in a 16 bit
value. This meant that the maximum altitude was 32 meters. Storing the
value to the precision sent from the Rx raised the limit to 328
meters, which isn't really enough.

Since it comes from the Rx in two 16-bit records, we just store both
of those and calculate the value for display when we display it
instead of when we receive it. The altitude display uses the standard
telemetry value display instead of the GPS altitude display, so no new
display handling code was needed.