Added extra channel for bayang protocol, from silver13
This is the "emergency stop channel" which works on the boldclash b-03 quad. Because of the altitude hold, this is needed if the quad fails to detect landing. Uses channel 12.

There are 2 other channels which are not implemented due to lack of room ( all 12 channels are used ) they are auto takeoff/land and "heavy lift mode" whatever that is, not as important as stop channel.

Tested on 7e

A new model file needs to be made because the number of channels is saved in it, and this is channel 12, it won't show otherwise.

The quad acts on channel change only.