CFlie: Adding 'simple' telemetry mode
Added 'simple' telemetry mode that reads RSSI and VBat (if present) out
of the NRF ACK packet. This will be faster and require less overhead
than establishing a full CRTP logging session, but cannot be extended
beyond RSSI/Internal VBat.

Also did a bit of refactoring of the CRTP logging code, and added RSSI
to the logging block.