Fix YD717 protocol binding when auto-ack is disabled, form hexfet
The Use A.ACK option was added recently to accommodate nRF clone chips that don't support auto-acknowledge correctly. This change missed updating the bind procedure, which uses auto-ack to skip binding if an already bound receiver is detected. The fix is to always go through the bind procedure when the option is set to No.