Support old format of localized file
Use the first character of second line to detect the version. All the old lang.* is starting with character ':'.
Load the old format also in the order of hash to support binary search.
Add a macro to wrapper the support to old format. target_defs.h can drop the support for old format to save some bytes in ROM.