Introduce GUI_Localize as string callback for label and button (#621)
* Introduce GUI_Localize as string callback for label and button

In the localize build, GUI_Localize is a wrapper to _tr. In non-localized build, it is NULL, which results the same behavior as today.

This also removes the assumption on that _tr returns a memory as static memory region. So this result will be used every time we need the localized string.