Implement CRSFshot synchronization between tx and radio.
Use CRSFshot "offset" telemetry to synchronize packet sends with the
transmitter module.  This reduces latency by delivering data to the
transmitter in sync with when the transmitter will use it.

Prior to this change a Crossfire TX LCD shows "CRSF V2" for the RC
Input.  After this change the display is "CRSFshot".